scripts/reference_data/requirements.txt
wheel
xmltodict
requestsw